pyproject.toml
[project]
name = "test-fixer"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Piotr Grędowski", email = "piotrgredowski@gmail.com" }]
dependencies = [
    "openai>=0.28.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.9.0",
    "pydantic>=2.3.0",
    "gitignore_parser>=0.1.6",
    "result>=0.13.1",
    "loguru>=0.7.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.2",
    "black>=23.9.1",
    "ruff>=0.0.289",
    "pytest-mock>=3.11.1",
    "ipython>=8.12.2",
]

[tool.rye.scripts]
test-examples = { cmd = "cd tests/functional/simple; python -m pytest examples" }
fix-tests = { }

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
markers = [
    "functional: mark test as functional",
    "example: mark test as example",
]
